A-Packages Moved all to package duke
