+++ src/main/java/duke/Deadline.java
@@ -1,6 +1,7 @@
+package duke;
+
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
+
 public class Deadline extends Task{
     protected String type = "[ D ]";
     protected LocalDate date;
