+++ src/main/java/duke/Event.java
@@ -1,5 +1,6 @@
+package duke;
+
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 
 public class Event extends Task{
     protected String type = "[ E ]";
